+++ src/settings/JobProfiles/JobProfileDetailsRoute/JobProfileDetailsRoute.js
@@ -43,6 +43,7 @@ const JobProfileDetailsRoute = ({
       isDefaultProfile={isDefaultProfile}
       isLoading={!jobProfileRecord || !mappingProfileRecord || (!isDefaultProfile && jobExecutionsQuery.isLoading)}
       onCancel={handleCancel}
+      onEdit={() => { history.push(`/settings/data-export/job-profiles/edit/${match.params.id}`); }}
       onDelete={() => DELETE({ id: jobProfileRecord?.id })}
     />
   );