+++ CHANGELOG.md
@@ -2,7 +2,8 @@
 
 ## IN PROGRESS
 
-*Select UUIDs type before starting export. UIDEXP-177
+* Select UUIDs type before starting export. UIDEXP-177
+* Use correct `css-loader` syntax. UIDEXP-255
 
 ## [5.0.0](https://github.com/folio-org/ui-data-export/tree/v4.0.0) (2021-10-08)
 [Full Changelog](https://github.com/folio-org/ui-data-export/compare/v4.1.0...v5.0.0)