+++ CHANGELOG.md
@@ -10,7 +10,7 @@
 * The list of errors is truncated. UIDEXP-259
 * Add total number of records to be exported to the Running component. UIDEXP-161
 * Refactor away from SafeHTMLMessage. UIDEXP-216
-
+* When go to Settings > Data export, change focus automatically to data-export pane header. UIDEXP-264
 
 ## [5.0.0](https://github.com/folio-org/ui-data-export/tree/v4.0.0) (2021-10-08)
 [Full Changelog](https://github.com/folio-org/ui-data-export/compare/v4.1.0...v5.0.0)